% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmst2perm.R
\name{print.rmst2perm}
\alias{print.rmst2perm}
\title{print.rmst2perm}
\usage{
\method{print}{rmst2perm}(x, digits = 3, ...)
}
\arguments{
\item{x}{Object to be printed}

\item{digits}{Integer indicating the number of decimal places}

\item{...}{Further arguments ignored in this function}
}
\value{
returns summary output for class 'rmst2perm'
}
\description{
S3 method for class 'rmst2perm'
}
